requirements.txt
Pillow
numpy
opencv-python
svgwrite